+++ client/pages/qna/answer/[id].tsx
@@ -1,34 +1,16 @@
 import { GetServerSideProps } from 'next';
 import { Params } from 'next/dist/shared/lib/router/utils/route-matcher';
-import dynamic from 'next/dynamic';
-
-import { useEffect } from 'react';
 
 import {
-  DehydratedState,
-  HydrationBoundary,
   QueryClient,
   dehydrate,
 } from '@tanstack/react-query';
 
-import styled from 'styled-components';
-import { styleMixin } from '@/src/shared/styles';
-
-import Spinner from '@/components/Spinner';
-
 import board from '@/src/entities/qnaboard/queryKeys';
-import { boardApi, useFeedStore } from '@/src/entities/qnaboard';
+import { boardApi } from '@/src/entities/qnaboard';
 
 import { getServerCookie } from '@/src/shared/utils';
 import { TOKEN_KEYS } from '@/src/shared/const';
-
-const AnswerLayout = dynamic(
-  () => import('@/components/QNA/Answer/AnswerLayout'),
-  {
-    ssr: false,
-    loading: () => <Spinner />,
-  },
-);
 
 // TODO:: 쿠키 파싱 처리하기
 export const getServerSideProps: GetServerSideProps = async (context) => {
@@ -48,32 +30,4 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
   };
 };
 
-type AnswerProps = {
-  dehydratedState: DehydratedState;
-};
-
-function Answer({ dehydratedState }: AnswerProps) {
-  const { initTargetFeed } = useFeedStore();
-  useEffect(
-    () => () => {
-      initTargetFeed();
-    },
-    [],
-  );
-
-  return (
-    <ViewStyle>
-      <HydrationBoundary state={dehydratedState}>
-        <AnswerLayout />
-      </HydrationBoundary>
-    </ViewStyle>
-  );
-}
-
-export default Answer;
-
-const ViewStyle = styled.div`
-  ${styleMixin.Column()}
-  width: 100%;
-  height: 100vh;
-`;
+export { default } from '@/src/pages/qna/answer';
