+++ client/src/features/boundary/index.ts
@@ -1,0 +1,1 @@
+export { default as APIErrorBoundary } from './APIErrorBoundary';
