+++ client/src/features/feedback/index.ts
@@ -1,0 +1,1 @@
+export { default as FeedBackComposer } from './FeedBackComposer';
