+++ client/src/shared/components/dropbox/DropMenuItem.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 
 import TestProvider from '@/fixutures/TestProvider';
 import DropMenuItem from '@/src/shared/components/dropbox/DropMenuItem';
@@ -13,6 +13,7 @@ function MockCompoent() {
 
 describe('DropMenuItem 컴포넌트', () => {
   const context = describe;
+  const mockAction = jest.fn();
 
   context('컴포넌트 렌더링될 때 type이 link 라면', () => {
     it('링크 아이템을 렌더링한다.', () => {
@@ -29,6 +30,7 @@ describe('DropMenuItem 컴포넌트', () => {
                 title: 'Test Title',
               },
             }}
+            offBox={mockAction}
           />
         </TestProvider>,
       );
@@ -36,6 +38,11 @@ describe('DropMenuItem 컴포넌트', () => {
       const link = screen.getByRole('link', { name: 'Test Label' });
       const title = screen.getByText('Test Title');
       const icon = screen.getByTestId('mock-icon');
+
+      fireEvent.click(link);
+      fireEvent.click(title);
+      fireEvent.click(icon);
+      expect(mockAction).toHaveBeenCalledTimes(3);
 
       expect(link).toBeInTheDocument();
       expect(title).toBeInTheDocument();
@@ -53,12 +60,16 @@ describe('DropMenuItem 컴포넌트', () => {
               type: 'component',
               component: <MockCompoent />,
             }}
+            offBox={mockAction}
           />
         </TestProvider>,
       );
 
       const isComponent = screen.getByText('컴포넌트');
       const icon = screen.queryByTestId('mock-icon');
+
+      fireEvent.click(isComponent);
+      expect(mockAction).toHaveBeenCalled();
       expect(icon).not.toBeInTheDocument();
       expect(isComponent).toBeInTheDocument();
     });
