+++ client/src/shared/components/spinner/index.ts
@@ -1,0 +1,1 @@
+export { default as Spinner } from './Spinner';
